backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
python-multipart==0.0.9
pymongo==4.6.3
python-dotenv==1.0.1
